requirements.txt
# Core dependencies
python-dotenv>=1.0.0
docarray>=0.32.0
numpy>=1.24.0
vectordb>=0.4.0

# AI and ML dependencies
semantic-kernel>=1.0.0
openai>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Optional web dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0

# Optional CLI dependencies
click>=8.1.0
rich>=13.0.0 